improved separation of gui & logic
