+++ modules/logic/brute_force.py
@@ -1,6 +1,7 @@
 from math import log2, ceil
 
 from modules.logic import legal
+from modules.gui import visualisation
 
 import __main__
 if "2d" in __main__.__file__:
@@ -8,29 +9,35 @@ if "2d" in __main__.__file__:
 else:
     from modules.gui import squares_3d as squares
 
+def solve(self):
+    visualisation.reset_board(self)
+    if self.candidates_found == True:
+        self.candidates = {pos: cand_list for pos, cand_list in sorted(self.candidates.items(), key= lambda item: len(item[1]))}
+    self.squares_list = self.grid_all_squares if self.candidates_found == False else list(self.candidates.keys())
+    
+    self.iterations = 0
+    return rec_solve(self, 0)
 
 def rec_solve(self, index):
-        
         # sudoku is solved when all squares are filled
         if index == len(self.squares_list):
             return True
         
         pos = self.squares_list[index]
-        i, j = pos[0], pos[1]
         
         # move onto next square if current square is already filled
-        if self.grid[i][j] != 0:            
+        if self.grid[pos[0]][pos[1]] != 0:            
             return rec_solve(self, index+1)
 
-        cands = range(1,10) if self.candidates_found == False else self.candidates[(i,j)]
+        cands = range(1,10) if self.candidates_found == False else self.candidates[pos]
         
         for num in cands:
             if legal.check(self, pos, num):
-                self.grid[i][j] = num
+                self.grid[pos[0]][pos[1]] = num
 
                 if self.var_show_iterating.get() == 1:
                     if self.candidates_found == True:
-                        squares.hide_candidates(self,pos)                    
+                        squares.hide_candidates(self, pos)                    
                         self.window.update()    
 
                     squares.set_value(self,pos,num)
@@ -44,11 +51,11 @@ def rec_solve(self, index):
                     return True
                 self.iterations += 1
 
-            self.grid[i][j] = 0
+            self.grid[pos[0]][pos[1]] = 0
 
             if self.var_show_iterating.get() == 1:
                 if self.candidates_found == True:
-                    squares.show_candidates(pos)
+                    squares.show_candidates(self, pos)
                     self.window.update()    
 
                 squares.set_value(self,pos,0)
